feat: 컴포넌트 분리 및 FSD에 따른 경로 정리 (#82)
