+++ src/shared/ui/text-editor/WordCount.tsx
