+++ src/shared/ui/text-editor/menuBarState.ts
